Crate sos_kernel

SOS kernel

This crate contains the kernel for SOS, the Stupid Operating System.

SOS: the Stupid Operating System

SOS is a simple, tiny toy OS implemented in Rust. It targets the x86, x86_64, and ARM v7 CPU architectures.

I'm writing this mostly for fun, to learn more about OS design and kernel hacking, so don't expect anything new or exciting out of this project.

SOS is copyright 2015-2016 Hawk Weisman, and is released under the terms of the MIT license.

Modules

arch
Architecture-specific implementation.

elf
Parsing and loading Executable and Linkable Format (ELF) 32- and 64-bit binaries.

io
Kernel IO.

memory
Kernel memory management.

multiboot2
Code for reading & extracting data from Multiboot 2 boot information.

panic
Panic handling.

util
General purpose stuff I couldn't find a better home for.

Functions

kernel_main
Kernel main loop

kernel_start
Kernel initialization function called from ASM