Module sos_kernel::arch

Architecture-specific implementation.

This module consists of a number of modules containing architecture-specific code for each targeted architecture. The arch module uses conditional compilation to re-export the implementation for which the kernel is currently being compiled.

In order for the rest of the kernel to work properly, an architecture-specific implementation module should define a number of specific items. If these are not defined, the platform-independant kernel implementation cannot function properly.

Please note that currently only the architecture-specific implementation for x86_64 (long mode) is implemented. The armv7 and x86 (protected mode) modules are currently much less complete.

Modules

bda
BIOS Data Area

drivers

interrupts

multiboot2
Code for reading & extracting data from Multiboot 2 boot information.

Constants

ARCH_BITS

Statics

HEAP_BASE

HEAP_TOP

STACK_BASE

STACK_TOP

Functions

arch_init
Entry point for architecture-specific kernel init

long_mode_init⚠
Trampoline to ensure we have a correct stack frame for calling arch_init