Module sos_kernel::arch::cpu

Code for interacting with the x86_64 CPU.

This module contains code for interrupts, paging, context switches, CPU I/O (from x86_all), and reading/writing the x86 control registers.

Modules

context
x86_64 execution contexts.

control_regs
x86 and x86_64 control registers

dtable
x86 and x86_64 descriptor tables (IDT, GDT, or LDT)

flags
Flags present in the %eflags/%rflags register on x86 CPUs.

gdt
The Global Descriptor Table (GDT) is used for configuring segmentation.

interrupts
64-bit Interrupt Descriptor Table implementation.

segment
Code for using the x86 and x86_64 segmentation hardware.

task
Tasking

timer

Structs

Port

Registers
Registers pushed to the stack when handling an interrupt or context switch.

Enums

PrivilegeLevel
Represents an x86 privilege level.