Trait sos_kernel::arch::cpu::dtable::DTable

pub trait DTable: Sized {
    type Entry: Sized;
    fn entry_count(&self) -> usize;
    unsafe fn load(&self);

    fn get_ptr(&self) -> Pointer { ... }
}

A descriptor table (IDT or GDT)

Associated Types

type Entry: Sized

Required Methods

fn entry_count(&self) -> usize

Returns the number of Entries in the DTable.

This is used for calculating the limit.

unsafe fn load(&self)

Load the descriptor table with the appropriate load instruction

Provided Methods

fn get_ptr(&self) -> Pointer

Get the IDT pointer struct to pass to lidt or lgdt

This expects that the object implementing DTable not contain additional data before or after the actual DTable, if you wish to attach information to a descriptor table besides the array of entries that it consists of, it will be necessary to encose the descriptor table in another struct or enum type.