Module sos_kernel::arch::cpu::interrupts

64-bit Interrupt Descriptor Table implementation.

Refer to section 6.10 of the Intel® 64 and IA-32 Architectures Software Developer’s Manual for more information. Intterupt handling on x86 machines.

This module provides support for interrupt handling on both x86 and x86_64 as a black box. Code that depends on this can use the same API regardless of system word size.

Modules

idt
Common functionality for the x86 and x86_64 Interrupt Descriptor Table.

pics
Support for the 8259 Programmable Interrupt Controller.

Structs

ExceptionInfo

InterruptContext
State stored when handling an interrupt.

Statics

EXCEPTIONS
x86 exceptions.

Functions

handle_cpu_exception
Handle a CPU exception with a given interrupt context.

handle_interrupt
Kernel interrupt-handling function.

initialize
Initialize interrupt handling.