Struct sos_kernel::arch::cpu::interrupts::idt::gate::Gate

pub struct Gate {
    pub offset_lower: u16,
    pub selector: Selector,
    pub flags: GateFlags,
    pub offset_mid: u16,
    pub offset_upper: u32,
    // some fields omitted
}

An IDT entry is called a gate.

Based on code from the OS Dev Wiki.

Refer also to "6.14.1 64-Bit Mode IDT" and "Table 3-2. System-Segment and Gate-Descriptor Types" in the Intel® 64 and IA-32 Architectures Software Developer’s Manual

Fields
offset_lower: u16

bits 0 - 15 of the offset

selector: Selector

code segment selector (GDT or LDT)

flags: GateFlags

indicates the gate's type and attributes. the second half indicates the type: + 0b1100: Call gate + 0b1110: Interrupt gate + 0b1111: Trap Gate

offset_mid: u16

bits 16 - 31 of the offset

offset_upper: u32

bits 32 - 63 of the offset

Methods

impl Gate

const fn absent() -> Self

Creates a new IDT gate marked as absent.

This is basically just for filling the new IDT table with valid (but useless) gates upon init.

Actually triggering an absent interrupt will send a General Protection fault (13).