Struct sos_kernel::arch::cpu::interrupts::idt::Idt

pub struct Idt(_);

An Interrupt Descriptor Table

The IDT is either 64-bit or 32-bit.

Methods

impl Idt

const fn new() -> Self

Construct a new IDT with all interrupt gates set to absent.

unsafe fn enable_interrupts()

Enable interrupts

unsafe fn disable_interrupts()

Disable interrupts

fn add_gate(&mut self, idx: usize, handler: Handler)

Add a new interrupt gate pointing to the given handler

unsafe fn add_handlers(&mut self) -> &mut Self

Add interrupt handlers exported by assembly to the IDT.