Module sos_kernel::arch::cpu::segment

Code for using the x86 and x86_64 segmentation hardware.

For more information, refer to the Intel® 64 and IA-32 Architectures Software Developer’s Manual, Vol. 3A, section 3.2, "Using Segments". Some of the documentation present in this module was taken from the Intel manual.

Structs

CodeFlags

DataFlags

Descriptor
A segment descriptor is an entry in an IDT or GDT.

Flags

Selector
A segment selector is a 16-bit identifier for a segment.

Enums

SysType

Type

Constants

ACCESSED
If this is a code or data segment and the accessed bit is set, it has been accessed.

AVAILABLE

CODE_ACCESSED

CODE_DATA_ACC
1 if this is a code or data segment that has been accessed

CONFORMING

DATA_ACCESSED

DEFAULT_SIZE

DESCR_TYPE

DPL

EXECUTE

EXEC_ONLY

EXPAND_DOWN

GRANULARITY

LENGTH

LIMIT

PRESENT

READ

RPL
Requested Prrivelege Level (RPL)

RPL_RING_0

RPL_RING_1

RPL_RING_2

RPL_RING_3

SEGMENT_TYPE

TI_GDT
If the Table Indicator (TI) is 0, use the GDT

TI_LDT
If the TI is 1, use the LDT

WRITE