Struct sos_kernel::arch::memory::paging::table::Table

pub struct Table<L> where L: TableLevel {
    pub entries: [Entry; 512],
    // some fields omitted
}

A page table

Fields
entries: [Entry; 512]

The entries in the page table.

Methods

impl<L: TableLevel> Table<L>

fn zero(&mut self)

Zeroes out the page table by setting all entries "unused"

impl<L: Sublevel> Table<L>

fn next_table(&self, index: usize) -> Option<&Table<L::Next>>

Returns the next table, or None if none exists

fn next_table_mut(&self, index: usize) -> Option<&mut Table<L::Next>>

Mutably borrows the next table.

fn create_next<A>(&mut self, index: usize, alloc: &mut A) -> &mut Table<L::Next> where A: FrameAllocator<Frame>

Returns the next table, creating it if it does not exist.