Struct sos_kernel::arch::multiboot2::Info

#[repr(C)]

pub struct Info {
    pub length: u32,
    // some fields omitted
}

Fields
length: u32
Methods

impl Info

fn start_addr(&self) -> PAddr

fn end_addr(&self) -> PAddr

unsafe fn from(addr: PAddr) -> Result<&'static Self, &'static str>

TODO: rewrite this as a TryFrom implementation (see issue #85)

fn get_tag(&'static self, tag_type: TagType) -> Option<&'static Tag>

Finds the tag with the given tag type.

This is actually safe since the tag types are constrained by The TagType enum

Returns

Some(tag) if a tag of the given type could be found.

None if no tag of the given type could be found.

fn mem_map(&'static self) -> Option<&'static MemMapTag>

Finds the memory map tag.

# Returns - Some(MemMapTag) if a memory map tag could be found - None if no tag of the given type could be found.

fn elf_sections(&'static self) -> Option<&'static ElfSectionsTag>

Finds the ELF sections tag.

# Returns - Some(ElfSectionsTag) if a memory map tag could be found - None if no tag of the given type could be found.

fn kernel_frames(&'static self) -> Result<FrameRange, &'static str>

Returns the kernel frame range from the Multiboot 2 ELF Sections