Struct sos_kernel::elf::file::Header

pub struct Header<W: ElfWord> {
    pub ident: Ident,
    pub machine: Machine,
    pub entry_point: W,
    pub ph_offset: W,
    pub sh_offset: W,
    pub flags: u32,
    pub header_size: u16,
    pub ph_entry_size: u16,
    pub ph_count: u16,
    pub sh_entry_size: u16,
    pub sh_count: u16,
    pub sh_str_idx: u16,
    // some fields omitted
}

An ELF file header

Fields
ident: Ident machine: Machine entry_point: W

Program entry point

ph_offset: W

Offset for start of program headers

sh_offset: W

Offset for start of section headers

flags: u32 header_size: u16 ph_entry_size: u16 ph_count: u16 sh_entry_size: u16 sh_count: u16 sh_str_idx: u16
Methods

impl<W> Header<W> where W: ElfWord

fn from_slice<'a>(input: &'a [u8]) -> ElfResult<&'a Header<W>>

Attempt to extract an ELF file header from a slice of bytes.

fn get_type(&self) -> Type

impl Header<u64>

fn parse_section<'a>(&'a self, input: &'a [u8], idx: u16) -> ElfResult<&'a Section>

Attempt to extract a section header from a slice of bytes.

impl Header<u32>

fn parse_section<'a>(&'a self, input: &'a [u8], idx: u16) -> ElfResult<&'a Section>

Attempt to extract a section header from a slice of bytes.