Enum sos_kernel::elf::section::Header

pub enum Header<'a> {
    ThirtyTwo(&'a HeaderRepr<u32>),
    SixtyFour(&'a HeaderRepr<u64>),
}

Represents an ELF section header

Refer to the ELF standard for more information.

Variants
ThirtyTwo(&'a HeaderRepr<u32>)SixtyFour(&'a HeaderRepr<u64>)
Methods

impl<'a> Header<'a>

fn offset(&self) -> u64

fn length(&self) -> u64

fn link(&self) -> u32

fn info(&self) -> u32

fn address_align(&self) -> u64

fn entry_length(&self) -> u64

fn addr(&self) -> PAddr

Returns the start address of this section

fn end_addr(&self) -> PAddr

Returns the end address of this section

fn flags(&self) -> Flags

Returns the type of this section

fn get_type(&self) -> ElfResult<Type>

Returns the type of this section

fn is_writable(&self) -> bool

Returns true if this section is writable.

fn is_allocated(&self) -> bool

Returns true if this section occupies memory during program execution.

fn is_executable(&self) -> bool

Returns true if this section contains executable instructions.

fn is_mergeable(&self) -> bool

Returns true if this section can be merged.

fn is_uniform(&self) -> bool

Returns true if this section contains data that is of a uniform size.