Trait sos_kernel::memory::Page

pub trait Page where Self: Sized, Self: AddAssign<usize> + SubAssign<usize>, Self: PartialEq<Self> + PartialOrd<Self>, Self: Copy + Clone {
    type Address: Addr<Self::R>;
    type R;
    fn containing(addr: Self::Address) -> Self;
    fn base(&self) -> Self::Address;

    fn range_between(start: Self, end: Self) -> PageRange<Self> { ... }
    fn range_until(&self, end: Self) -> PageRange<Self> { ... }
}

Trait for a page. These can be virtual pages or physical frames.

Associated Types

type Address: Addr<Self::R>

The type of address used to address this Page.

Typically, this is a PAddr or VAddr (but it could be a "MAddr") in schemes where we differentiate between physical and machine addresses. If we ever have those.

type R

Required Methods

fn containing(addr: Self::Address) -> Self

Returns a new Page containing the given Address.

N.B. that since trait functions cannot be const, implementors of this trait may wish to provide implementations of this function outside of the impl block and then wrap them here.

fn base(&self) -> Self::Address

Returns the base address where this page starts.

Provided Methods

fn range_between(start: Self, end: Self) -> PageRange<Self>

Returns a PageRange between two pages

fn range_until(&self, end: Self) -> PageRange<Self>

Returns a FrameRange on the frames from this frame until the end frame